Refactor to use async-http-client
Generally, folks in the Swift sswg (Server Side Working Group) recommend
using Swift NIO as a foundational toolkit for networking, and built out
the [AsyncHTTPClient](https://github.com/swift-server/async-http-client)
leveraging that and many best practices.

This is a heavily breaking change, moving away from SwiftyJSON in favor
of built-in support for Codable, as well as the AsyncHTTPClient since
that should be well-supported and able to place nice with other packages
across the ecosystem.